+++ unifi_console/CHANGELOG.md
@@ -1,2 +1,15 @@
 # CHANGELOG - Unifi Console
 
+
+## 1.2.0
+* [Added] Add support for device specific metrics (UAP, UDM, USG, USW, USX). See [#1715](https://github.com/DataDog/integrations-extras/pull/1715)
+* [Fix] Fix checks pilling up
+
+## 1.1.1
+* [Fix] Fix url issue for non UMP setup. See [#1596](https://github.com/DataDog/integrations-extras/pull/1596)
+
+## 1.1.0
+* [Added] Add support for UMP and update dashboard. See [#1562](https://github.com/DataDog/integrations-extras/pull/1562)
+
+## 1.0.0
+* Unifi initial realease. See [#1351](https://github.com/DataDog/integrations-extras/pull/1351)